SPAPP-4: Allow to import data from github advisory
  * Continue work on github deserialization
