+++ application-security-advisory-github/pom.xml
@@ -53,5 +53,11 @@
       <groupId>com.fasterxml.jackson.core</groupId>
       <artifactId>jackson-databind</artifactId>
     </dependency>
+    <dependency>
+      <groupId>org.xwiki.commons</groupId>
+      <artifactId>xwiki-commons-tool-test-component</artifactId>
+      <scope>test</scope>
+      <version>${commons.version}</version>
+    </dependency>
   </dependencies>
 </project>
