+++ application-security-advisory-github/src/main/resources/META-INF/components.txt
@@ -1,0 +1,2 @@
+org.xwiki.contrib.securityadvisory.internal.github.GithubAdvisoryDeserializer
+org.xwiki.contrib.securityadvisory.internal.github.GithubImporter
