+++ application-security-advisory-github/src/test/java/org/xwiki/contrib/securityadvisory/internal/github/GithubImporterTest.java
@@ -1,0 +1,15 @@
+package org.xwiki.contrib.securityadvisory.internal.github;
+
+import org.xwiki.test.junit5.mockito.ComponentTest;
+import org.xwiki.test.junit5.mockito.InjectMockComponents;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+@ComponentTest
+class GithubImporterTest
+{
+    @InjectMockComponents
+    private GithubImporter importer;
+
+
+}
